Use double frametime for timestamps
Yay, no more jitter! Doubles for compatibility, because floats are too inaccurate when timestamps are millisecond-based.

+++ GhostReplay/ScriptSettings.hpp
@@ -23,7 +23,7 @@ public:
 
     struct {
         // Normal
-        float OffsetSeconds = 0.0f;
+        double OffsetSeconds = 0.0;
         int VehicleAlpha = 40;
         int ForceLights = 0;
